requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PCA9685==1.0.1
Adafruit-PureIO==1.1.9
click==8.1.3
colorama==0.4.5
dominate==2.6.0
Flask==2.1.3
Flask-Bootstrap==3.3.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyYAML==6.0
visitor==0.1.3
Werkzeug==2.1.2
